requirements.txt
# requirements.txt

# Web 框架
fastapi>=0.110.0
uvicorn[standard]>=0.29.0

# 数据验证
pydantic>=2.0.0

# 配置
pyyaml>=6.0

# 数据库
sqlalchemy>=2.0.0
pymysql>=1.1.0       # MySQL 驱动，sqlite 不需要

# Elasticsearch
elasticsearch>=8.0.0,<9.0.0

# 模型推理
torch>=2.0.0
transformers>=4.40.0
sentence-transformers>=2.7.0

# PDF 解析
pdfplumber>=0.10.0
pymupdf>=1.24.0      # 备用解析器，处理复杂 PDF

# LLM 调用
openai>=1.0.0

# 评估（第四阶段加）
# ragas>=0.1.0

# 测试
pytest>=8.0.0
httpx>=0.27.0        # FastAPI TestClient 依赖